Move global states to a dedicated Context file
