+++ src/styles/components/App.module.css
@@ -4,44 +4,3 @@
   align-items: center;
   min-height: 100vh;
 }
-
-.appContainer > main {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-  padding: 30px 20px 50px;
-  flex: 1;
-}
-
-.appContainer > main > h1 {
-  font-size: 30px;
-  font-weight: 700;
-}
-
-.postsContainer {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(280px, 350px));
-  justify-items: center;
-  justify-content: space-evenly;
-  gap: 50px 30px;
-
-  width: 100%;
-  margin: 30px 0 50px;
-}
-
-.categoriesContainer {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  gap: 12px;
-
-  width: 100%;
-  margin: 30px 0 60px;
-}
-
-@media (min-width: 800px) {
-  .appContainer > main {
-    padding: 50px 80px;
-  }
-}
